HexagonalArrays

Documentation for HexagonalArrays.

HexagonalArrays.HexagonalIndex

HexagonalArrays.heuclidean

HexagonalArrays.to_cartesian

HexagonalArrays.HexagonalIndex — Type

HexagonalIndex(i, j, k)   -> I
HexagonalIndex((i, j, k)) -> I

Create an hexagonal index I in Hexagonal Efficient Coordinate System (HECS), which is used for indexing an hexagonal array A.

In HECS a point on an hexagonal grid is represented by three coordinates $(i, j, k) \in \mathbb{Z}^{*} \times \mathbb{Z}^{*} \times \{1, 2\}$. You should note that the original HECS indexing is adapted to the Julia indexing system (column major and 1-based indexing). Hence, $i$ corresponds to the row number, $j$ to the column number and $k$ to the column parity (i.e. when $k = 1$ you access to an odd column, and when $k = 2$ you access to an even column).

In HECS space four operators are defined for vector arithmetic :

addition $I_1 + I_2$

negation $-I$

subtraction $I_1 - I_2 = I_1 + (-I_2)$

scalar multiplication $s * I, s \in \mathbb{N}$ and $-s * I = s * (-I)$

The Julia operators +, - and * are extended for these four operations.

Examples

julia> I1 = HexagonalIndex(1, 2, 1)
HexagonalIndex((1, 2, 1))

julia> I2 = HexagonalIndex(3, 1, 1)
HexagonalIndex((3, 1, 1))

julia> I1 + I2
HexagonalIndex((4, 3, 1))

julia> I1 - I2
HexagonalIndex((-2, 1, 1))

julia> 3 * I1
HexagonalIndex((3, 6, 1))

HexagonalArrays.heuclidean — Method

heuclidean(I1, I2, unitd)

Return the euclidean distance between two hexagonal indices I1 and I2.

See also to_cartesian.

HexagonalArrays.to_cartesian — Method

to_cartesian(I, unitd) -> (x, y)

Return the cartesian coordinates (x, y) of the hexagonal index I. The upper-left hexagon of the hexagonal grid is assumed to have cartesian coordinates (0,0). unitd is the unit distance between two adjacent hexagons.